+++ packages/liveblocks-core/src/__tests__/LiveMap.test.ts
@@ -2,10 +2,10 @@ import { RoomScope } from "../AuthToken";
 import { LiveList } from "../LiveList";
 import { LiveMap } from "../LiveMap";
 import { LiveObject } from "../LiveObject";
-import { WebsocketCloseCodes } from "../types";
 import { OpCode } from "../types/Op";
 import type { IdTuple, SerializedCrdt } from "../types/SerializedCrdt";
 import { CrdtType } from "../types/SerializedCrdt";
+import { WebsocketCloseCodes } from "../types/Websocket";
 import {
   createSerializedList,
   createSerializedMap,
