+++ packages/liveblocks-core/src/__tests__/LiveObject.test.ts
@@ -1,10 +1,10 @@
 import { LiveList } from "..";
 import { RoomScope } from "../AuthToken";
 import { LiveObject } from "../LiveObject";
-import { WebsocketCloseCodes } from "../types";
 import { OpCode } from "../types/Op";
 import type { IdTuple, SerializedCrdt } from "../types/SerializedCrdt";
 import { CrdtType } from "../types/SerializedCrdt";
+import { WebsocketCloseCodes } from "../types/Websocket";
 import {
   createSerializedList,
   createSerializedObject,
