+++ packages/liveblocks-core/src/types/Websocket.ts
@@ -1,0 +1,11 @@
+export enum WebsocketCloseCodes {
+  CLOSE_ABNORMAL = 1006,
+
+  INVALID_MESSAGE_FORMAT = 4000,
+  NOT_ALLOWED = 4001,
+  MAX_NUMBER_OF_MESSAGES_PER_SECONDS = 4002,
+  MAX_NUMBER_OF_CONCURRENT_CONNECTIONS = 4003,
+  MAX_NUMBER_OF_MESSAGES_PER_DAY_PER_APP = 4004,
+  MAX_NUMBER_OF_CONCURRENT_CONNECTIONS_PER_ROOM = 4005,
+  CLOSE_WITHOUT_RETRY = 4999,
+}
